+++ app/src/main/AndroidManifest.xml
@@ -11,13 +11,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <service
-            android:name=".PlaySoundService"
-            android:enabled="true"
-            android:exported="false" />
         <receiver android:name=".AlarmReceiver"/>
 
-        <activity android:name="com.example.alarmapp.PlaySoundActivity" />
         <activity
             android:name=".Title"
             android:label="@string/app_name" />
