+++ app/src/main/java/com/example/alarmapp/Home.java
@@ -5,12 +5,14 @@ import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
 import android.media.MediaPlayer;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 
 public class Home extends AppCompatActivity {
 
     private MediaPlayer mediaPlayer;
+    private float volume = 0.3f;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -22,7 +24,17 @@ public class Home extends AppCompatActivity {
 
         if(flag){
             mediaPlayer = MediaPlayer.create(this, R.raw.sample);
+            mediaPlayer.setVolume(volume, volume);
             mediaPlayer.start();
+            mediaPlayer.setOnCompletionListener(new MediaPlayer.OnCompletionListener() {
+                @Override
+                public void onCompletion(MediaPlayer mediaPlayer) {
+                    Log.d("onCompletion", "=== active ===");
+                    volume += 0.1f;
+                    mediaPlayer.setVolume(volume, volume);
+                    mediaPlayer.start();
+                }
+            });
         }
 
         Button timeSet_button = findViewById(R.id.timeSet_button);
@@ -38,7 +50,8 @@ public class Home extends AppCompatActivity {
         stop_button.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                stopService(new Intent(Home.this, PlaySoundService.class));
+                stopAlarm();
+                // FIXME === cloud store に起床したことを表す信号を投げる ===
 
                 Intent intent_stop = new Intent(getApplication(),Stay.class);
                 startActivity(intent_stop);
@@ -46,5 +59,14 @@ public class Home extends AppCompatActivity {
         });
     }
 
+    private void stopAlarm() {
+        if(mediaPlayer != null) {
+            mediaPlayer.stop();
+            mediaPlayer.release();
+            mediaPlayer = null;
+        }
+    }
+
 
 }
+
